+++ code/src/cljs/sixsq/nuvla/ui/main/events.cljs
@@ -189,10 +189,11 @@
                            #(dispatch [::set-bootsrap-message %])]}))
 
 
-(reg-event-db
-  ::set-welcome-message
-  (fn [db [_ welcome-message]]
-    (assoc db ::spec/welcome-message welcome-message)))
+(reg-event-fx
+  ::set-message
+  (fn [{db :db} [_ type message]]
+    (cond-> {:db (assoc db ::spec/message [type, message])}
+            message (assoc :dispatch-later [{:ms 10000 :dispatch [::set-message nil]}]))))
 
 
 (reg-event-db