+++ code/src/cljs/sixsq/nuvla/ui/welcome/views.cljs
@@ -4,6 +4,7 @@
     [re-frame.core :refer [dispatch subscribe]]
     [sixsq.nuvla.ui.authn.events :as authn-events]
     [sixsq.nuvla.ui.authn.subs :as authn-subs]
+    [sixsq.nuvla.ui.authn.utils :as authn-utils]
     [sixsq.nuvla.ui.history.events :as history-events]
     [sixsq.nuvla.ui.i18n.subs :as i18n-subs]
     [sixsq.nuvla.ui.main.subs :as main-subs]
@@ -31,16 +32,19 @@
   [path]
   (let [tr           (subscribe [::i18n-subs/tr])
         iframe?      (subscribe [::main-subs/iframe?])
-        query-params (subscribe [::main-subs/nav-query-params])]
+        query-params (subscribe [::main-subs/nav-query-params])
+        {:keys [reset-password, message, error]} @query-params]
     (when @query-params
 
-      (when (contains? @query-params :message)
-        (dispatch [:sixsq.nuvla.ui.main.events/set-welcome-message (-> @query-params :message keyword)]))
+      (when (or message error)
+        (dispatch [:sixsq.nuvla.ui.main.events/set-message
+                   (if error :error :success) (or error message)]))
 
-      (when (contains? @query-params :reset-password)
-        (dispatch [::authn-events/set-form-id "session-template/password-reset"])
+      (when reset-password
+        (dispatch [::authn-events/set-form-id authn-utils/session-tmpl-password-reset])
         (dispatch [::authn-events/update-form-data :username (:reset-password @query-params)])
         (dispatch [::authn-events/open-modal :reset-password]))
+
       (dispatch [::history-events/navigate (str (first path) "/")]))
     [ui/Container {:textAlign "center"
                    :fluid     true